+++ packages/expo-dev-menu/android/src/debug/java/expo/modules/devmenu/DevMenuHost.kt
@@ -10,10 +10,10 @@ import com.facebook.react.ReactPackage
 import com.facebook.react.ReactPackageTurboModuleManagerDelegate
 import com.facebook.react.bridge.JavaScriptExecutorFactory
 import com.facebook.react.config.ReactFeatureFlags
+import com.facebook.react.devsupport.DevMenuReactInternalSettings
 import com.facebook.react.devsupport.DevServerHelper
 import com.facebook.react.shell.MainReactPackage
 import devmenu.com.th3rdwave.safeareacontext.SafeAreaContextPackage
-import expo.modules.devmenu.react.DevMenuReactInternalSettings
 import expo.modules.devmenu.react.createNonDebuggableJavaScriptExecutorFactory
 import java.io.BufferedReader
 import java.io.FileNotFoundException